refactor: separate web-view components into individual files
- Extracted CommandCard component from command-list.tsx (83 lines)
- Extracted GroupEditDialog component from group-command-editor.tsx (74 lines)
- Updated import paths to use new separated components
- Followed React principle of one exported component per file

fix #83
